backend/requirements.txt
aiofiles==23.2.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
bidict==0.23.1
billiard==4.2.1
boto3==1.34.14
botocore==1.34.162
cachetools==5.5.0
celery==5.3.6
certifi==2024.8.30
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.5
Django==4.2.16
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-environ==0.8.1
django-filter==23.5
django-phonenumber-field==8.0.0
django-rest-swagger==2.2.0
django-storages==1.14.2
django-timezone-field==7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
ecdsa==0.19.0
fastapi==0.115.5
ffmpy==0.4.0
filelock==3.16.1
fsspec==2024.10.0
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.66.0
gradio==5.6.0
gradio_client==1.4.3
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jmespath==1.0.1
kombu==5.4.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgpack==1.1.0
numpy==2.1.3
oauthlib==3.2.2
openapi-codec==1.3.2
orjson==3.10.11
packaging==24.2
pandas==2.2.3
pillow==10.2.0
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.28.3
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.2.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.5
python-dotenv==1.0.1
python-engineio==4.10.1
python-jose==3.3.0
python-multipart==0.0.12
python-socketio==5.10.0
pytz==2024.2
PyYAML==6.0.2
redis==5.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
ruff==0.7.4
s3transfer==0.10.3
safehttpx==0.1.1
semantic-version==2.10.0
setuptools==75.5.0
shellingham==1.5.4
simple-websocket==1.1.0
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.2
starlette==0.41.2
tomlkit==0.12.0
tqdm==4.67.0
typer==0.13.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
vine==5.1.0
wcwidth==0.2.13
websockets==12.0
wsproto==1.2.0
zoom-python==0.1.1
